\name{auxfunctions-package}
\alias{auxfunctions-package}
\alias{auxfunctions}
\docType{package}
\title{
\packageTitle{auxfunctions}
}
\description{
\packageDescription{auxfunctions}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{auxfunctions}

Maintainer: \packageMaintainer{auxfunctions}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = auxfunctions)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{auxfunctions}}
%\section{Documentation Index}{\packageIndices{auxfunctions}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
