% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{Format_Integer_LeadingZeros}
\alias{Format_Integer_LeadingZeros}
\title{Reformat integers to strings with leading zeroes}
\usage{
Format_Integer_LeadingZeros(
  myInteger,
  myWidth = 2,
  myPrefix = "",
  mySuffix = ""
)
}
\arguments{
\item{myInteger}{a vector with numerical values}

\item{myWidth}{a numerical value, indicating the width of the transformed number
(number of digits)}

\item{myPrefix}{an optional character value or vector of the same length as myInteger}

\item{mySuffix}{an optional character value or vector of the same length as myInteger}
}
\value{
a vector of character strings
}
\description{
Format_Integer_LeadingZeros converts a number or a vector of numbers to a character format
with leading zeros. Optionally prefixes
}
\examples{
demovector_values <- c (1, 2, 3, 4, 5)
demo_prefix <- "ID."
demo_suffix <- ".TA"
Format_Integer_LeadingZeros (
  myInteger = demovector_values,
  myWidth   = 4,
  myPrefix  = demo_prefix,
  mySuffix  = demo_suffix
)
## Result:  "ID.0001.TA" "ID.0002.TA" "ID.0003.TA" "ID.0004.TA" "ID.0005.TA"
}
