% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{Format_Numeric_Percentage}
\alias{Format_Numeric_Percentage}
\title{Reformat integers to strings in a percentage format}
\usage{
Format_Numeric_Percentage(myNumeric, myDigits = 0, myReplacementNA = 0)
}
\arguments{
\item{myNumeric}{a vector with numerical values}

\item{myDigits}{a numerical value, indicating the decimal places}

\item{myReplacementNA}{an optional value to be provided when the input is NA}
}
\value{
a vector of character strings
}
\description{
Format_Numeric_Percentage converts a number or a vector of numbers to character strings
with percentage format.
}
\examples{
demovector_1 <- c (0.11, 0.82, 0.987654, 0.5, 0.51, NA)
Format_Numeric_Percentage (
  myNumeric  = demovector_1,
  myDigits   = 1,
  myReplacementNA = -9999
)
## Result:  "11.0 \%"    "82.0 \%"    "98.8 \%"    "50.0 \%"    "51.0 \%"    "-9999.0 \%"
}
