% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{Parse_DataFrame_Variable_Index}
\alias{Parse_DataFrame_Variable_Index}
\title{Simplify the use of variable names with an index as suffix}
\usage{
Parse_DataFrame_Variable_Index(
  myDataFrame,
  myVariableName,
  myIndex,
  myDigits = 2
)
}
\arguments{
\item{myDataFrame}{a dataframe}

\item{myVariableName}{a part of the variable name not including the index}

\item{myIndex}{an integer value or a vector of integer values}

\item{myDigits}{an integer value indicating the number of decimal places of the result}
}
\value{
A numerical value or a vector of numerical values
}
\description{
Parse_DataFrame_Variable_Index is applied to dataframes with column names
that have an index as suffix and simplifies extracting parts of it
(for example in loops).
}
\examples{

  Parse_DataFrame_Variable_Index(
   myDataFrame = clidamonger::data.ta.hd [101:120,],
   myVariableName = "M_2021_",
   myIndex = 1:12,
   myDigits = 1)

## Result:  0.4  3.4   NA  3.5  3.6  3.1   NA   NA   NA   NA -1.8   NA  2.9  3.4  2.6  2.5  4.4   NA   NA   NA
#  These are the values of the datasets number 101 to 120
#  and the months 1 to 12 of the year 2021 (columns "M_2021_01", "M2021_02", ... "M2021_12")

}
