% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{Parse_StringAsList}
\alias{Parse_StringAsList}
\title{Convert a character string with specific separators to a list}
\usage{
Parse_StringAsList(myString, mySeparator = ",", myStringToNA = "_NA_")
}
\arguments{
\item{myString}{the input string}

\item{mySeparator}{a character used as separator}

\item{myStringToNA}{a character string that will be converted to NA, except when set to FALSE}
}
\value{
A vector with character strings
}
\description{
Parse_StringAsList evaluates a character string containing specific separators
and returns a vector of the character strings betweeen the seperators
}
\examples{
demostring_1 <- "A,B,_NA_,C"
Parse_StringAsList (
  myString = demostring_1,
  mySeparator = ","
)
## Result: "A" "B" NA  "C"

demostring_2 <- "A|B|C|D|X|F|G"
Parse_StringAsList (
  myString     = demostring_2,
  mySeparator  = '|',
  myStringToNA = "X"
)
## Result: "A" "B" "C" "D" NA  "F" "G"

}
