% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{Reformat_InputData_Boolean}
\alias{Reformat_InputData_Boolean}
\title{Reformat input data to 0/1-boolean}
\usage{
Reformat_InputData_Boolean(myIndicator)
}
\arguments{
\item{myIndicator}{a vector with strings, numbers or boolean}
}
\value{
the integer values 1 or 0
}
\description{
Reformat_InputData_Boolean converts TRUE and "1" to 1 as well as FALSE and "0" to 0
}
\examples{
temp1 <- c (FALSE, TRUE, "0", "1", 0, 1, "Anything", 137, NA)
Reformat_InputData_Boolean (myIndicator = temp1)
## Result:  0  1  0  1  0  1 NA NA NA
}
