% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{Replace_NA}
\alias{Replace_NA}
\title{Replace NA (Excel equivalent IFERROR)}
\usage{
Replace_NA(myValue, ReturnIfNA)
}
\arguments{
\item{myValue}{the vector with values to be checked}

\item{ReturnIfNA}{the output value to be used in case of NA}
}
\value{
the input vector with NA values converted
}
\description{
Replace_NA replaces NA or Inf by a value or a string.
The function is used to simplify parsing Excel formulas
}
\examples{
temp1 <- c (1, 2, 3, NA, 5, 6)
Replace_NA (myValue = temp1, ReturnIfNA = "MissingValue")
## Result: "1"            "2"            "3"            "MissingValue" "5"            "6"
}
