% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{Replace_NULL}
\alias{Replace_NULL}
\title{Replace NULL (replace not existing element of a dataframe)}
\usage{
Replace_NULL(myValue, ReturnIfNULL)
}
\arguments{
\item{myValue}{the variable of a dataframe that might not yet exist}

\item{ReturnIfNA}{the output value to be used in case of NULL}
}
\value{
the input vector if existing or the ReturnIfNA value if not existing
}
\description{
Replace_NULL can be used in an operation on an element of a dataframe, that might not yet be existing,
and thus can help to avoid a result NULL.
NULL might be replaced by a value, a string or by NA.
An existing element or vector will not be changed.
}
\examples{
temp1 <- as.data.frame (c ("a", "b", "c", "d"))

colnames (temp1) <- "Letters"

Replace_NULL (temp1$Letters, NA)
## Result: "a" "b" "c" "d"

Replace_NULL (temp1$Numbers, NA)
## Result: NA

paste0 (temp1$Letters, Replace_NULL (temp1$Numbers, 0))
## Result: "a0" "b0" "c0" "d0"

Replace_NA (1/0, NA)
## Result NA

}
