% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{String_ParTab}
\alias{String_ParTab}
\title{Provide a character string from a parameter table}
\usage{
String_ParTab(
  myParTab,
  myCode,
  myDataFieldName,
  myCodeForNoComponent = "-",
  myValueForNoComponent = "",
  myErrorValue = "_ERROR_"
)
}
\arguments{
\item{myParTab}{the table to be evaluated}

\item{myCode}{a character string identifying the parameter ID = first column of the parameter table}

\item{myDataFieldName}{a character string identifying the datafield name (variable name)
= first row of the parameter table}

\item{myCodeForNoComponent}{an optional code identifying a row of the parameter table
to be used when the component is not existing}

\item{myValueForNoComponent}{an optional string used when the component is not existing}

\item{myErrorValue}{= an optional error string used when the element of the table cannot be identified}
}
\value{
A character string
}
\description{
String_ParTab provides a character string from a table or dataframe identified by the row name and the column name.
The function is similar to Value_ParTab but adapted to provide character strings instead of numerical values.
}
\examples{
temp_StationName <-
  String_ParTab (
    myParTab = clidamonger::data.ta.hd,
    myCode = "DE.MET.000917.TA_12",
    myDataFieldName = "Name_Station",
    myCodeForNoComponent = "-",
  )
temp_StationName
## Result: "Darmstadt"

}
