% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{TimeStampForDataset}
\alias{TimeStampForDataset}
\title{Time stamp for dataset (current date and time)}
\usage{
TimeStampForDataset()
}
\arguments{
\item{(None)}{No input}
}
\value{
A string containing separators "-", " " and ":"
}
\description{
TimeStampForDataset creates a string from current date and time in a standard date-time format.
}
\examples{
TimeStampForDataset ()
}
