% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{TimeStampForFileName}
\alias{TimeStampForFileName}
\title{Time stamp for filename (current date and time)}
\usage{
TimeStampForFileName()
}
\arguments{
\item{(none)}{no input}
}
\value{
a string containing separators "-" and "_"
}
\description{
TimeStampForFileName creates a string from current date and time in a format usable as a file name.
}
\examples{
TimeStampForFileName ()
}
