% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{Value_ParTab}
\alias{Value_ParTab}
\title{Provide a value from a parameter table}
\usage{
Value_ParTab(
  myParTab,
  myCode,
  myDataFieldName,
  myRoundDigits = 10,
  myCodeForNoComponent = "-",
  myValueForNoComponent = 0,
  myErrorValue = -99999
)
}
\arguments{
\item{myParTab}{the table to be evaluated}

\item{myCode}{a character string identifying the parameter ID = first column of the parameter table}

\item{myDataFieldName}{a character string identifying the datafield name (variable name)
= first row of the parameter table}

\item{myRoundDigits}{an optional integer indicating the number of digits for rounding the result value}

\item{myCodeForNoComponent}{an optional code identifying a row of the parameter table
to be used when the component is not existing}

\item{myValueForNoComponent}{an optional value used when the component is not existing}

\item{myErrorValue}{= an optional error value used when the element of the table cannot be identified}
}
\value{
A numerical or string value
}
\description{
Value_ParTab provides a value from a table or dataframe identified by the row name and the column name.
}
\examples{
temp_StationName <-
  Value_ParTab (
    myParTab = clidamonger::data.ta.hd,
    myCode = "DE.MET.000917.TA_12",
    myDataFieldName = "Name_Station",
    myCodeForNoComponent = "-",
  )
temp_StationName
## Result: "Darmstadt"

temp_TA_12_M_2021_12 <-
  Value_ParTab (
    myParTab = clidamonger::data.ta.hd,
    myCode = "DE.MET.000917.TA_12",
    myDataFieldName = "M_2021_12",
    myRoundDigits = 1,
    myCodeForNoComponent = "-",
    myValueForNoComponent = -99,
    myErrorValue = -99
  )
temp_TA_12_M_2021_12
}
