\name{write_to_1_flat_file}
\alias{write_to_1_flat_file}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Write One File to a Flat File
}
\description{
This function is designed to combine the content of multiple files into one flat file, along with headers indicating the file names.}
\usage{
write_to_1_flat_file(flat_file_name, file_name_to_write)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{flat_file_name}{The name of the flat file where the content will be written.}

  \item{file_name_to_write}{The name of the file whose content will be written to the flat file.}
}
\details{
Here's the breakdown of the `write_to_1_flat_file` function:

1. **Function Definition:**
   - `write_to_1_flat_file <- function (flat_file_name, file_name_to_write)`: Defines a function named `write_to_1_flat_file` that takes two arguments:
     - `flat_file_name`: The name of the flat file where the content will be written.
     - `file_name_to_write`: The name of the file whose content will be written to the flat file.

2. **Redirecting Output:**
   - `sink(flat_file_name, append = TRUE)`: Redirects the output to the specified `flat_file_name` using `sink()`. The `append = TRUE` argument ensures that the content is appended to the file if it already exists.

3. **Reading File Content:**
   - `current_file_text <- readr::read_lines(file_name_to_write)`: Reads the content of the file specified by `file_name_to_write` using `readr::read_lines()` and stores it in the variable `current_file_text`.

4. **Printing Header:**
   - `cat("\\n\\n#### Current file:", basename(file_name_to_write), "----\\n\\n")`: Prints a header indicating the current file being written to the flat file. `basename(file_name_to_write)` extracts the base name of the file.

5. **Writing File Content:**
   - `cat(current_file_text, sep = "\\n")`: Writes the content of `current_file_text` to the flat file using `cat()`. The `sep = "\\n"` argument ensures that each line is separated by a newline character.

6. **Restoring Output:**
   - `sink()`: Closes the output redirection, restoring the standard output behavior.
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Anna Shipunova
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
