% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{xl_AND}
\alias{xl_AND}
\title{Logical AND (Excel equivalent AND)}
\usage{
xl_AND(MyVector_1, MyVector_2, MyVector_3 = TRUE, MyVector_4 = TRUE)
}
\arguments{
\item{myVector_1}{the first vector containing boolean values}

\item{myVector_2}{the second vector containing boolean values, same length as myVector_1}

\item{myVector_3}{an optional third vector containing boolean values, same length as myVector_1}

\item{myVector_4}{an optional fourth vector containing boolean values, same length as myVector_1}
}
\value{
a vector with boolean values
}
\description{
xl_AND evaluates boolean values or vectors of boolean values
0 and FALSE as well as 1 and TRUE are treated in the same way.
The function is used to simplify parsing Excel formulas.
}
\examples{
temp1 <- c (TRUE, TRUE, FALSE, FALSE, TRUE)
temp2 <- c (TRUE, FALSE, TRUE, FALSE, TRUE)
temp3 <- c (0, 1, 0, 0, 1)
xl_AND (temp1, temp2, temp3)
# Result: FALSE FALSE FALSE FALSE  TRUE
}
