% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{xl_AVERAGE}
\alias{xl_AVERAGE}
\title{Average vectors by row  (Excel equivalent AVERAGE)}
\usage{
xl_AVERAGE(
  myVector_01,
  myVector_02,
  myVector_03 = NA,
  myVector_04 = NA,
  myVector_05 = NA,
  myVector_06 = NA,
  myVector_07 = NA,
  myVector_08 = NA,
  myVector_09 = NA,
  myVector_10 = NA
)
}
\arguments{
\item{myVector_01}{the first vector containing numeric values}

\item{myVector_02}{the second vector containing numeric values}

\item{myVector_03}{to myVector_10 optional further vectors containing numeric values}
}
\value{
a vector with character strings
}
\description{
xl_AVERAGE evaluates numeric vectors and
provides a vector of mean values of the same length as the input vectors.
The function is used to simplify parsing Excel formulas and apply it to vector variables.
}
\examples{
temp1 <- c (0, 1, 2, 3, 4)
temp2 <- c (0.5, 1.5, 2.5, 3.5, 4.5)
xl_AVERAGE (temp1, temp2)
# Result:  "000" "001" "002" "003" "004"

}
