% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{xl_CONCATENATE}
\alias{xl_CONCATENATE}
\title{Concatenate vectors of character strings by row (Excel equivalent CONCATENATE)}
\usage{
xl_CONCATENATE(..., collapse = NULL, recycle0 = FALSE)
}
\arguments{
\item{...}{one or more R objects, to be converted to character vectors.}
}
\value{
A character vector of the concatenated values.
}
\description{
xl_CONCATENATE concatenates vectors of character strings by row.
The function is used to simplify parsing Excel formulas and to apply them to vector variables.
}
\examples{
temp1 <- c ("cow ", "pig ", "elephant ", "tiger ")
temp2 <- c ("eats ", "loves ", "grows ", "has ")
temp3 <- c ("grass", "mud", "big", "stripes")
xl_CONCATENATE (temp1, temp2, temp3)
# Result: "cow eats grass"     "pig loves mud"      "elephant grows big" "tiger has has "
}
