% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{xl_ConvertDate}
\alias{xl_ConvertDate}
\title{Convert Excel date from integer to date format}
\usage{
xl_ConvertDate(x, origin = "1900-01-01", ...)
}
\arguments{
\item{x}{a value or vector of integers}

\item{origin}{a string representing the origin of the Excel date, default = "1900-01-01"}
}
\value{
a value or vector of R date format
}
\description{
xl_ConvertDate converts dates imported from Excel by openxlsx and thus only existing as integer to an R date format
}
\examples{
xl_ConvertDate (c (35968, 35440))
}
