% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{xl_DATE}
\alias{xl_DATE}
\title{Provide date format from year, month and day integers  (Excel equivalent DATE)}
\usage{
xl_DATE(myYear = 1900, myMonth = 1, myDay = 1)
}
\arguments{
\item{myVector_01}{the first vector containing numeric values}

\item{myVector_02}{the second vector containing numeric values}

\item{myVector_03}{to myVector_10 optional the first vector containing numeric values}
}
\value{
a vector with character strings in date format
}
\description{
xl_DATE evaluates integer vectors containing numerical information about year, month and day
and provides a vector of dates in the format YYYY-MM-DD.
The function is used to simplify parsing Excel formulas and apply it to vector variables.
}
\examples{
temp_year  <- c (2020, 2021, 2021, 2022)
temp_month <- c (12, 1, 12, 1)
temp_day   <- c (31, 1, 31, 1)
xl_DATE (temp_year, temp_month, temp_day)
# Result:   "2020-12-31" "2021-01-01" "2021-12-31" "2022-01-01"

}
