% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{xl_EOMONTH}
\alias{xl_EOMONTH}
\title{Provide the last day of months  (Excel equivalent EOMONTH)}
\usage{
xl_EOMONTH(myDate = "1900-01-01", myAddMonths = 0)
}
\arguments{
\item{myDate}{a value or a vector containing the date}

\item{myAddMonths}{a value or a vector containing the number of additional months}
}
\value{
a vector with character strings in date format
}
\description{
xl_EOMONTH evaluates a vector with dates and provides for each date
the last day of the given month. In addition an integer number can be used to add
a certain number months to the given dates.
The function is used to simplify parsing Excel formulas and apply it to vector variables.
}
\examples{

temp_date_1  <- c ("2020-01-01", "2021-01-01", "2022-01-01", "2023-01-01", "2024-01-01")
xl_EOMONTH (temp_date_1, 0)
# Result: "2020-01-31" "2021-01-31" "2022-01-31" "2023-01-31" "2024-01-31"
xl_EOMONTH (temp_date_1, 1)
# Result: "2020-02-29" "2021-02-28" "2022-02-28" "2023-02-28" "2024-02-29"
temp_date_2  <- "1999-12-31"
temp_add_months <- c (2, 12+2, 2*12+2, 3*12+2, 4*12+2, 5*12+2)
xl_EOMONTH (temp_date_2, temp_add_months)
# Result: "2000-02-29" "2001-02-28" "2002-02-28" "2003-02-28" "2004-02-29" "2005-02-28"

}
