% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{xl_FIND}
\alias{xl_FIND}
\title{Middle section of strings (Excel equivalent MID)}
\usage{
xl_FIND(myPattern, myText, myCharIndex_Start = 1)
}
\arguments{
\item{myPattern}{the character string to be found}

\item{myText}{a string or a vector of strings to be evaluated}

\item{myCharIndex_Start}{the position in the strings to start the search}
}
\value{
a number or a vector of numbers giving the position in the string.
If the pattern is not found, the value -1 will be returned
}
\description{
xl_FIND provides the position of a of character strings in a vector, starting from a specific position
and consisting of an indicated number of characters.
The function is used to simplify parsing Excel formulas
}
\examples{
temp1 <- c ("2000-01-01 00:00", "2000-01-02 00:00", "2000-01-03 00:00", "2000-01-04 00:00", "2000-01-05 00:00")
xl_FIND (myPattern = "04", temp1, 1)
## Result: -1 -1 -1  9 -1

DF_Test <- as.data.frame (c("Test234",
                            "alskjfa34_kl",
                            "Test"),
                          row.names = c("Row1", "Row2", "Row3"))
colnames (DF_Test) <- "Text"
DF_Test$Pattern <-  c("34", "skj", "est")
DF_Test
# printed dataframe "DF_Test":
#              Text Pattern
# Row1      Test234      34
# Row2 alskjfa34_kl     skj
# Row3         Test     est
xl_FIND (DF_Test$Pattern, DF_Test$Text, 1)
# Result: 6 3 2

}
