% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{xl_LEFT}
\alias{xl_LEFT}
\title{Left side of strings (Excel equivalent LEFT)}
\usage{
xl_LEFT(myString, i_Char_LeftSide)
}
\arguments{
\item{myString}{a string or a vector of strings to be evaluated}

\item{i_Char_LeftSide}{the number of characters to be provided}
}
\value{
a string or a vector of strings consisting of the defined section of the input
}
\description{
xl_LEFT provides the left side of character strings in a vector, consisting of an indicated number of characters.
The function is used to simplify parsing Excel formulas
}
\examples{
temp1 <- c ("sixty", "sixty one", "sixty two", "sixty three")
xl_LEFT (temp1, 3)
## Result: "six" "six" "six" "six"

}
