% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{xl_MID}
\alias{xl_MID}
\title{Middle section of strings  (Excel equivalent MID)}
\usage{
xl_MID(myString, myCharIndex_Start, myCharCount)
}
\arguments{
\item{myString}{a string or a vector of strings to be evaluated}

\item{myCharIndex_Start}{the position to start the section}

\item{myCharCount}{the number of characters to be provided}
}
\value{
A string or a vector of strings consisting of the defined section of the input
}
\description{
xl_MID provides the middle section of character strings in a vector, starting from a specific position
and consisting of an indicated number of characters.
The function is used to simplify parsing Excel formulas
German Excel syntax: "TEIL ()"
}
\examples{
temp1 <- c ("2000-01-01 00:00", "2000-01-02 00:00", "2000-01-03 00:00", "2000-01-04 00:00", "2000-01-05 00:00")
xl_MID (temp1, 9, 2)
## Result: "01" "02" "03" "04" "05"

}
