% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{xl_NOT}
\alias{xl_NOT}
\title{Logical NOT (Excel equivalent NOT)}
\usage{
xl_NOT(myVector)
}
\arguments{
\item{myVector}{the vector containing boolean values}
}
\value{
a vector with boolean values
}
\description{
xl_OR evaluates a boolean value or a vector of boolean values
0 and FALSE as well as 1 and TRUE are treated in the same way.
The function is used to simplify parsing Excel formulas
}
\examples{
temp1 <- c (TRUE, TRUE, FALSE, FALSE, TRUE)
xl_NOT (temp1)
# Result: FALSE FALSE  TRUE  TRUE FALSE

temp2 <- c (0, 1, 0, 0, 1)
xl_NOT (temp2)
# Result:  TRUE FALSE  TRUE  TRUE FALSE

}
