% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{xl_OR}
\alias{xl_OR}
\title{Logical OR (Excel equivalent OR)}
\usage{
xl_OR(myVector_1, myVector_2, myVector_3 = FALSE, myVector_4 = FALSE)
}
\arguments{
\item{myVector_1}{the first vector containing boolean values}

\item{myVector_2}{the second vector containing boolean values, same length as myVector_1}

\item{myVector_3}{an optional third vector containing boolean values, same length as myVector_1}

\item{myVector_4}{an optional fourth vector containing boolean values, same length as myVector_1}
}
\value{
a vector with boolean values
}
\description{
xl_OR evaluates boolean values or vectors of boolean values
0 and FALSE as well as 1 and TRUE are treated in the same way.
The function is used to simplify parsing Excel formulas
}
\examples{
temp1 <- c (TRUE, TRUE, FALSE, FALSE, TRUE)
temp2 <- c (TRUE, FALSE, TRUE, FALSE, TRUE)
temp3 <- c (0, 1, 0, 0, 1)
xl_OR (temp1, temp2, temp3)
# Result:  TRUE  TRUE  TRUE FALSE  TRUE
}
