% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{xl_RIGHT}
\alias{xl_RIGHT}
\title{Right side of strings (Excel equivalent RIGHT)}
\usage{
xl_RIGHT(myString, i_Char_RightSide)
}
\arguments{
\item{myString}{a string or a vector of strings to be evaluated}

\item{i_Char_RightSide}{the number of characters to be provided}
}
\value{
A string or a vector of strings consisting of the defined section of the input
}
\description{
xl_RIGHT provides the right side of character strings in a vector, consisting of an indicated number of characters.
The function is used to simplify parsing Excel formulas
}
\examples{
temp1 <- c ("2000", "2001", "2002", "2003", "2004")
xl_RIGHT (temp1, 2)
## Result: "00" "01" "02" "03" "04"

}
