% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{xl_SUMPRODUCT}
\alias{xl_SUMPRODUCT}
\title{Sum of the elements of a vector product (Excel equivalent SUMPRODUCT)}
\usage{
xl_SUMPRODUCT(
  myVector_01,
  myVector_02,
  myVector_03 = NA,
  myVector_04 = NA,
  myVector_05 = NA,
  myVector_06 = NA,
  myVector_07 = NA,
  myVector_08 = NA,
  myVector_09 = NA,
  myVector_10 = NA
)
}
\arguments{
\item{myVector_01}{the first vector containing numeric values}

\item{myVector_02}{the second vector containing numeric values}

\item{myVector_03}{to myVector_10 optional: further vectors containing numeric values}
}
\value{
A numerical value
}
\description{
xl_SUMPRODUCT evaluates numeric vectors of the same length by multiplying
the elements of different vectors row by row and summing up the result vector
The function is used to simplify parsing Excel formulas and apply it to vector variables.
}
\examples{
temp1 <- c (0, 1, 2, 3, 4)
temp2 <- c (0, 1, 2, 3, 4)
temp3 <- c (0, 1, 2, 3, 4)
xl_SUMPRODUCT (temp1, temp2, temp3)
## Result:  100

}
