\name{see_res_in_outfile}
\alias{see_res_in_outfile}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Show the Long Text in an Outfile
}
\description{
This function provides a convenient way to visualize text data by saving it to a temporary text file and automatically opening it for viewing. It is particularly useful for displaying large or complex text outputs that may be difficult to analyze directly in the console.
}
\usage{
see_res_in_outfile(text_to_output)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{text_to_output}{
  string
}
}
\details{
This function, `see_res_in_outfile`, is designed to display text content in an external text file for easy viewing. Here's a detailed breakdown of its functionality:

1. **Creating a Temporary File:**
   - `outfile <- tempfile(fileext = ".txt")`: It generates a temporary file with the extension ".txt" using the `tempfile` function. The path to this file is stored in the variable `outfile`.

2. **Writing Text Content to File:**
   - `cat(text_to_output, file = outfile)`: The `cat` function is used to write the content of the `text_to_output` parameter to the file specified by `outfile`. This effectively saves the provided text content to the temporary file.

3. **Displaying the File:**
   - `file.show(outfile)`: This command opens the temporary file in the default text editor of the system, allowing the user to view the contents of the file.
}
\value{
Instead of an output shows a temp file.
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  Anna Shipunova
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
