\name{see_res_in_outfile}
\alias{see_res_in_outfile}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Show the Long Text in an Outfile
}
\description{
This function provides a convenient way to visualize text data by saving it to a temporary text file and automatically opening it for viewing. It is particularly useful for displaying large or complex text outputs that may be difficult to analyze directly in the console.
}
\usage{
see_res_in_outfile(text_to_output)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{text_to_output}{
  string or dataframe
}
}
\details{
- The `see_res_in_outfile` function is designed to output the content of a variable to a temporary text file and then display that file.

- `tempfile(fileext = ".txt")` creates a temporary file with the extension ".txt" and assigns its path to the variable `outfile`.

- `sink(file = outfile)` redirects R output to the file specified by `outfile`, meaning subsequent output will be written to this file.

- `print(var_to_output, na.print = "", n = nrow(var_to_output))` writes the content of `var_to_output` to the temporary file. The `na.print = ""` argument specifies that NA values should be printed as an empty string, and `n = nrow(var_to_output)` ensures all rows of the data frame are printed.

- `file.show(outfile)` opens the temporary file in the default text editor for viewing.

- `sink()` restores the normal R output to the console, stopping the redirection of output to the temporary file.}
\value{
Instead of an output shows a temp file.
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  Anna Shipunova
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
